Commenting, operator overloading and error feedback

+++ FinancialTrackerProj/Expense.h
@@ -6,11 +6,11 @@ class Expense : public Transaction
 {
 public:
 	Expense();
-	Expense(string description, int amount,string date);
+	Expense(string description, double amount,string date);
 	~Expense()=default;
 	string toString()const;
 	void print()const ;
-	int getAmount()const;
+	double getAmount()const;
 private:
 	string type;
 };
