+++ FinancialTrackerProj/Income.h
@@ -6,11 +6,11 @@ class Income : public Transaction
 {
 public:
 	Income();
-	Income(string description, int amount, string date);
+	Income(string description, double amount, string date);
 	~Income()=default;
 	string toString()const;
 	void print()const;
-	int getAmount()const;
+	double getAmount()const;
 private:
 	string type;
 };
